+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.4] - 2025-08-02
+
+### Added
+
+- Added a mechanism for matching an incoming set of parameters with a list of invalid values;
+- Added the definition of the most appropriate schema version;
+- Added unit tests for new functions;
+
+### Updated
+
+- Updated the function for collecting unique keys from an array of validation schemes for a set of parameters;
+
 ## [1.0.0-beta.3] - 2025-07-24
 
 ### Added